% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_desc.R
\name{reorder_desc}
\alias{reorder_desc}
\title{Reorder factor levels in descending order}
\usage{
reorder_desc(d)
}
\arguments{
\item{d}{factor}
}
\value{
factor
}
\description{
Reorder an exisiting factor levels in descending order, where the new factor levels are arranged in opposite orientation to the character readout
}
\examples{
#' reorder_desc(d)
}

