% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderv2-package.r
\docType{package}
\name{reorderv2}
\alias{reorderv2}
\alias{reorderv2-package}
\title{reorderv2, reorder in descending order}
\description{
Graphing factors tend to be messy, since they are usually misaligned.
Reorder is a function used to order factors in ascending order.
Reorderv2 is a package which will order factors in descending order.
}

